requirements.txt
# Required for AbuseIPDB and AlienVault API calls
requests
azure-functions
dnspython